Refactor

+++ physics-common/src/main/java/me/furkandgn/physicsdemo/common/Point.java
@@ -1,8 +1,8 @@
 package me.furkandgn.physicsdemo.common;
 
-public record Point(float x, float y) {
+public record Point(double x, double y) {
 
-  public static Point of(float x, float y) {
+  public static Point of(double x, double y) {
     return new Point(x, y);
   }
 }
